[FIX]던전씬 버그 일부 고침

+++ WeGoMars/WeGoMars/Character.cs
@@ -36,7 +36,7 @@ namespace WeGoMars{
 
 		public abstract float Attack();
 
-		public abstract void TakeDamage(float damage);
+		public abstract void TakeDamage(int damage);
 
 		public Item DropItem()
 		{
